+++ types/failures/failures_test.go
@@ -1,6 +1,7 @@
 package failures
 
 import (
+	"os"
 	"sort"
 	"testing"
 	"time"
@@ -13,8 +14,19 @@ import (
 )
 
 func TestMain(m *testing.M) {
+	tmpDir, err := os.MkdirTemp("", "statping-failures-test")
+	if err != nil {
+		os.Exit(1)
+	}
+	utils.Directory = tmpDir
+	utils.InitEnvs()
+	utils.Params.Set("STATPING_DIR", tmpDir)
 	_ = utils.InitLogs()
-	m.Run()
+
+	code := m.Run()
+
+	_ = os.RemoveAll(tmpDir)
+	os.Exit(code)
 }
 
 // mockColumnID implements ColumnIDInterfacer for testing